+++ modules/Food_Service/Students/Statements.php
@@ -170,7 +170,7 @@ if ( UserStudentID() && ! $_REQUEST['modfunc'] )
 
 			$link['remove']['link'] = PreparePHP_SELF(
 				$_REQUEST,
-				[ 'delete_cancel' ],
+				[],
 				[ 'modfunc' => 'delete' ]
 			);