+++ modules/Food_Service/Users/Statements.php
@@ -117,7 +117,7 @@ if ( UserStaffID() && ! $_REQUEST['modfunc'] )
 
 			$link['remove']['link'] = PreparePHP_SELF(
 				$_REQUEST,
-				[ 'delete_cancel' ],
+				[],
 				[ 'modfunc' => 'delete' ]
 			);